Typo in Sync._refresh_views (#331)
Just upgraded to 2.3.2 and got the following fatal error:
```
Exception in poll_redis() for thread Thread-94: 'Sync' object has no attribute 'materialized_views'
```
This was the only place in the code that matched.  It looks like this is a simple typo in converting from `self.materialized` to `self._materialized_views`.  There should be an underscore.  Thank you for developing this library.